Split GraphValidationTest into separate test classes, as well as moving some tests to ComponentValidationTest.
RELNOTES=n/a

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=209591715
